na float verandert

+++ src/main/java/test/Test.java
@@ -7,8 +7,8 @@ public class Test {
 		double x= 100;
 		int y = 2;
 		for(int i = 0; i < 100; i++){
-			double f =(y/x)*((y+x)*y*55.55555) ;
-			f = (int) (f * 100+ 0.5) / 100;
+			float f =(float) ((y/x)*((y+x)*y*55.55555)) ;
+			f =  (float) ((f * 100+ 0.5) / 100);
 			System.out.println("lvl: " + (i+1) + " -> "+ f);
 			x = f;
 			y += 1;
